refactor: reorganize API paths
