+++ convex/user/fetchLogs.ts
@@ -1,5 +1,5 @@
 import { v } from "convex/values";
-import { query } from "./_generated/server";
+import { query } from "../_generated/server";
 
 export const fetchLogs = query({
   args: {
