+++ convex/user/submitLog.ts
@@ -1,5 +1,5 @@
 import { v } from "convex/values";
-import { mutation } from "./_generated/server";
+import { mutation } from "../_generated/server";
 
 export const submitLog = mutation({
   args: {
